# swaggerPublic.json
{
  "swagger": "2.0",
  "info": {
    "title": "Subscription Public API",
    "description": "Subscription public API is a microservice which communicates with Subsciption Service. ",
    "version": "1.0.0"
  },
  "host": "localhost:undefined",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "api-key",
      "in": "header"
    }
  },
  "paths": {
    "/bff/getSubscription": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/bff/createSubscription": {
      "post": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/bff/deleteSubscription": {
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api-docs": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {}
      }
    }
  },
  "definitions": {
    "server side error": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "ERROR"
        },
        "msg": {
          "type": "string",
          "example": "some error message"
        },
        "error": {
          "type": "object",
          "properties": {
            "message": {
              "type": "string",
              "example": "Error message caught"
            },
            "name": {
              "type": "string",
              "example": "Error name"
            },
            "stack": {
              "type": "string",
              "example": "Error stack"
            }
          },
          "required": [
            "message",
            "name"
          ]
        }
      },
      "required": [
        "status",
        "msg"
      ]
    },
    "calculation": {
      "type": "object",
      "properties": {
        "createdAt": {
          "type": "string",
          "example": "2020-03-31T00:00:00.000Z"
        },
        "result": {
          "type": "number",
          "example": 100
        }
      },
      "required": [
        "createdAt",
        "result"
      ]
    }
  }
}